requirements.txt
openai>=1.0.0,<2.0.0
sentence-transformers>=3.0.0,<4.0.0
chromadb>=0.5.0,<1.0.0
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
torch>=2.6.0,<3.0.0
langchain>=0.3.0,<1.0.0
langchain-openai>=0.2.0,<1.0.0
langchain-community>=0.3.0,<1.0.0
